+++ src/dataregistry/DataRegistry.py
@@ -18,6 +18,7 @@ class DataRegistry:
         root_dir=None,
         verbose=False,
         site=None,
+        production_mode=False,
     ):
         """
         Primary data registry wrapper class.
@@ -54,18 +55,20 @@ class DataRegistry:
             Can be used instead of `root_dir`. Some predefined "sites" are
             built in, such as "nersc", which will set the `root_dir` to the
             data registry's default data location at NERSC.
+        production_mode : bool, optional
+            True to register/modify production schema entries
         """
 
         # Establish connection to database
-        self.db_connection = DbConnection(config_file, schema=schema,
-                                          verbose=verbose)
+        self.db_connection = DbConnection(
+            config_file, schema=schema, verbose=verbose, production_mode=production_mode
+        )
 
         # Work out the location of the root directory
         self.root_dir = self._get_root_dir(root_dir, site)
 
         # Create registrar object
-        self.Registrar = Registrar(self.db_connection, self.root_dir,
-                                   owner, owner_type)
+        self.Registrar = Registrar(self.db_connection, self.root_dir, owner, owner_type)
 
         # Create query object
         self.Query = Query(self.db_connection, self.root_dir)