+++ src/dataregistry/registrar/dataset.py
@@ -143,11 +143,11 @@ class DatasetTable(BaseTable):
         if kwargs_dict["owner_type"] == "production":
             if kwargs_dict["is_overwritable"]:
                 raise ValueError("Cannot overwrite production entries")
-            if (not self._table_metadata.is_production_schema) and (
+            if (not self.db_connection.production_mode) and (
                 not kwargs_dict["test_production"]
             ):
                 raise ValueError(
-                    "Only the production schema can handle owner_type='production'"
+                    "Must be in `production_mode` to write to production schema'"
                 )
 
             # The only owner allowed for production datasets is "production"
@@ -155,7 +155,7 @@ class DatasetTable(BaseTable):
                 raise ValueError("`owner` for production datasets must be 'production'")
         else:
             if self._dialect != "sqlite" and not kwargs_dict["test_production"]:
-                if self._table_metadata.is_production_schema:
+                if self.db_connection.production_mode:
                     raise ValueError(
                         "Only owner_type='production' can go in the production schema"
                     )
@@ -833,9 +833,7 @@ class DatasetTable(BaseTable):
         """
 
         # Find the dataset entry with this combination
-        previous = self._find_previous(
-            name, version_string, owner, owner_type
-        )
+        previous = self._find_previous(name, version_string, owner, owner_type)
 
         if len(previous) == 0:
             raise ValueError(
@@ -1016,9 +1014,10 @@ class DatasetTable(BaseTable):
                 )
 
                 result = conn.execute(stmt)
+                rows = result.fetchall()
 
                 # If we don't have the keyword, add it
-                if result.rowcount == 0:
+                if len(rows) == 0:
                     add_table_row(
                         conn,
                         dataset_keyword_table,